CREATING SERVICE BEFORE SOURCE
There is an existing issue with the eventing controller that
sometimes prevents the PingSource from reconcile. This will
eventually cause our source to fail with SinkNotFound exception.
In this case, we need to have the Sink/Service in place before
we create the Source.

More details on #2994
Reference: https://github.com/knative/eventing/issues/2994